Add jacoco report
