+++ src/test/java/FelineTest.java
@@ -9,6 +9,7 @@ import org.mockito.Spy;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.times;
 
 @RunWith(MockitoJUnitRunner.class)
 public class FelineTest {
@@ -19,10 +20,12 @@ public class FelineTest {
     @Test
     public void getFamilyTest()  {
         assertEquals("Кошачьи", feline.getFamily());
+        Mockito.verify(feline, times(1)).getFamily();
     }
 
     @Test
     public void getKittensTest() {
         assertEquals(1, feline.getKittens());
+        Mockito.verify(feline, times(1)).getKittens();
     }
 }
